backend/requirements.txt
flask
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-vertexai
langchain-pinecone
pinecone-client
python-dotenv
google.generativeai
langchain_groq
langgraph
playwright